Migrate Swipe Controller to it's own file
